fix a signed/unsigned mismatch (/W4 C4389)
git-svn-id: https://mpc-hc.svn.sourceforge.net/svnroot/mpc-hc/trunk@5164 10f7b99b-c216-0410-bff0-8a66a9350fd8

+++ src/Subtitles/RTS.cpp
@@ -682,7 +682,7 @@ bool CPolygon::ParseStr()
 
 bool CPolygon::CreatePath()
 {
-    size_t len = m_pathTypesOrg.GetCount();
+    int len = (int)m_pathTypesOrg.GetCount();
     if (len == 0) {
         return false;
     }
@@ -693,7 +693,7 @@ bool CPolygon::CreatePath()
         if (!mpPathTypes || !mpPathPoints) {
             return false;
         }
-        mPathPoints = (int)len;
+        mPathPoints = len;
     }
 
     memcpy(mpPathTypes, m_pathTypesOrg.GetData(), len * sizeof(BYTE));
